+++ Domibus-MSH/src/main/java/eu/domibus/core/user/UserBase.java
@@ -26,7 +26,7 @@ public interface UserBase {
         }
     }
 
-    int getEntityId();
+    long getEntityId();
 
     UserBase.Type getType();
