+++ Domibus-MSH/src/main/java/eu/domibus/web/rest/ro/PluginUserRO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
  */
 public class PluginUserRO {
 
-    private Integer entityId;
+    private Long entityId;
 
     @SizeIfPresent(min = 4, max = 255)
     @Pattern(regexp = "^[a-zA-Z0-9\\.@_]*$")
@@ -43,11 +43,11 @@ public class PluginUserRO {
 
     private LocalDateTime expirationDate;
 
-    public Integer getEntityId() {
+    public Long getEntityId() {
         return entityId;
     }
 
-    public void setEntityId(Integer entityId) {
+    public void setEntityId(Long entityId) {
         this.entityId = entityId;
     }
